chore: tsify browser-window
