+++ spec-main/ambient.d.ts
@@ -19,21 +19,6 @@ declare namespace Electron {
   interface Session {
     destroy(): void;
   }
-
-  // Experimental views API
-  class TopLevelWindow {
-    constructor(args: {show: boolean})
-    setContentView(view: View): void
-  }
-  class WebContentsView {
-    constructor(options: BrowserWindowConstructorOptions)
-  }
-
-  namespace Main {
-    class TopLevelWindow extends Electron.TopLevelWindow {}
-    class View extends Electron.View {}
-    class WebContentsView extends Electron.WebContentsView {}
-  }
 }
 
 declare module 'dbus-native';
